Article 8 Implementation/requirements.txt
# Core dependencies
tensorflow>=2.10.0
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.25.0

# Data processing
opencv-python>=4.7.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
imbalanced-learn>=0.10.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
pillow>=9.3.0

# Web application
flask>=2.2.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.28.0

# Kaggle API (for automatic dataset download)
kaggle>=1.5.12

tensorflow-hub
psutil
tensorflow_addons

einops

albumentations